# Improve url_decode performance for long strings