… cpp/benchmarks/CMakeLists.txt
… cpp/tests/strings/urls_tests.cpp