Fixes #7348.

This changes the `url_decode` algorithm from row-level parallelism to character-level parallelism which improves the performance when operating on string columns that have longer average string lengths.  A benchmark for `url_decode` has also been added.