@jlowe 2021-02-09
Using the benchmark, here are the before and after performance numbers on a V100.  The two numbers varying in the benchmark name are the number of rows and characters per row, respectively.
Before:
```
------------------------------------------------------------------------------------------------------------------
Benchmark                                                        Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------------
UrlDecode<10>/url_decode_10pct/100/10/manual_time            0.053 ms        0.071 ms        12289 bytes_per_second=25.0311M/s
UrlDecode<10>/url_decode_10pct/1000/10/manual_time           0.073 ms        0.091 ms         9156 bytes_per_second=181.77M/s
UrlDecode<10>/url_decode_10pct/10000/10/manual_time          0.077 ms        0.095 ms         8789 bytes_per_second=1.6881G/s
UrlDecode<10>/url_decode_10pct/100000/10/manual_time         0.084 ms        0.102 ms         8133 bytes_per_second=15.5132G/s
UrlDecode<10>/url_decode_10pct/100/100/manual_time           0.164 ms        0.182 ms         4217 bytes_per_second=60.2996M/s
UrlDecode<10>/url_decode_10pct/1000/100/manual_time          0.450 ms        0.468 ms         1554 bytes_per_second=220.233M/s
UrlDecode<10>/url_decode_10pct/10000/100/manual_time         0.451 ms        0.468 ms         1552 bytes_per_second=2.14842G/s
UrlDecode<10>/url_decode_10pct/100000/100/manual_time        0.789 ms        0.806 ms          885 bytes_per_second=12.2702G/s
UrlDecode<10>/url_decode_10pct/100/1000/manual_time           1.48 ms         1.50 ms          471 bytes_per_second=64.5066M/s
UrlDecode<10>/url_decode_10pct/1000/1000/manual_time          4.77 ms         4.79 ms          147 bytes_per_second=200.825M/s
UrlDecode<10>/url_decode_10pct/10000/1000/manual_time         4.91 ms         4.92 ms          143 bytes_per_second=1.90587G/s
UrlDecode<10>/url_decode_10pct/100000/1000/manual_time        33.9 ms         33.9 ms           21 bytes_per_second=2.76001G/s
UrlDecode<10>/url_decode_10pct/100/10000/manual_time          14.6 ms         14.6 ms           48 bytes_per_second=65.2909M/s
UrlDecode<10>/url_decode_10pct/1000/10000/manual_time         49.3 ms         49.3 ms           14 bytes_per_second=193.504M/s
UrlDecode<10>/url_decode_10pct/10000/10000/manual_time        49.7 ms         49.7 ms           14 bytes_per_second=1.87403G/s
UrlDecode<10>/url_decode_10pct/100000/10000/manual_time        371 ms          371 ms            2 bytes_per_second=2.5111G/s
```

After:
```
------------------------------------------------------------------------------------------------------------------
Benchmark                                                        Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------------
UrlDecode<10>/url_decode_10pct/100/10/manual_time            0.070 ms        0.087 ms         9599 bytes_per_second=19.1569M/s
UrlDecode<10>/url_decode_10pct/1000/10/manual_time           0.086 ms        0.104 ms         7839 bytes_per_second=155.34M/s
UrlDecode<10>/url_decode_10pct/10000/10/manual_time          0.103 ms        0.120 ms         6597 bytes_per_second=1.26665G/s
UrlDecode<10>/url_decode_10pct/100000/10/manual_time         0.208 ms        0.224 ms         3342 bytes_per_second=6.27197G/s
UrlDecode<10>/url_decode_10pct/100/100/manual_time           0.078 ms        0.095 ms         8643 bytes_per_second=127.449M/s
UrlDecode<10>/url_decode_10pct/1000/100/manual_time          0.094 ms        0.112 ms         7130 bytes_per_second=1051.62M/s
UrlDecode<10>/url_decode_10pct/10000/100/manual_time         0.190 ms        0.206 ms         3685 bytes_per_second=5.10589G/s
UrlDecode<10>/url_decode_10pct/100000/100/manual_time         1.27 ms         1.28 ms          553 bytes_per_second=7.65637G/s
UrlDecode<10>/url_decode_10pct/100/1000/manual_time          0.086 ms        0.103 ms         7868 bytes_per_second=1112.93M/s
UrlDecode<10>/url_decode_10pct/1000/1000/manual_time         0.173 ms        0.189 ms         4024 bytes_per_second=5.4189G/s
UrlDecode<10>/url_decode_10pct/10000/1000/manual_time         1.12 ms         1.14 ms          626 bytes_per_second=8.34177G/s
UrlDecode<10>/url_decode_10pct/100000/1000/manual_time        11.8 ms         11.8 ms           59 bytes_per_second=7.95281G/s
UrlDecode<10>/url_decode_10pct/100/10000/manual_time         0.153 ms        0.169 ms         4564 bytes_per_second=6.10873G/s
UrlDecode<10>/url_decode_10pct/1000/10000/manual_time         1.01 ms         1.02 ms          696 bytes_per_second=9.26655G/s
UrlDecode<10>/url_decode_10pct/10000/10000/manual_time        10.7 ms         10.7 ms           65 bytes_per_second=8.72123G/s
UrlDecode<10>/url_decode_10pct/100000/10000/manual_time        121 ms          121 ms            6 bytes_per_second=7.67838G/s

```